+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 
-## [5.38.0](https://github.com/plivo/plivo-dotnet/tree/v5.38.0) (2023-11-03)
+## [5.39.0](https://github.com/plivo/plivo-dotnet/tree/v5.39.0) (2023-11-09)
 **Feature - registration_status field**
 - Added new param `registration_status` in LIST Campaign APIs
+
+## [5.38.0](https://github.com/plivo/plivo-dotnet/tree/v5.38.0) (2023-11-08)
+**Verify Service**
+-Added Support for Verify Service Public API'S 
+-Create Session API(POST): To create a verify session (2FA). Allowed params recipient, method, channel, callback_url, app_uuid
+-Get Session API(GET): To retrieve a particular session
+-List Sessions : To retrieve all the sessions
+-Validate Session : To validate OTP for a particular session
 
 ## [5.37.0](https://github.com/plivo/plivo-dotnet/tree/v5.37.0) (2023-10-31)
 **Feature - TollFree Verification API Support**